[Refactor] Rename IndexCacheInterface to IndexCache and concrete to DefaultIndexCache
